# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VBUY — Premium Ghanaian E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/admins:
    get:
      operationId: listAdmins
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Admin list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
    post:
      operationId: inviteAdmin
      tags: [admin]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, roleId]
              properties:
                email: { type: string }
                roleId: { type: integer }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/admins/{id}:
    patch:
      operationId: updateAdmin
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                roleId: { type: integer }
                status: { type: string, enum: [active, disabled] }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AdminUser:
      type: object
      properties:
        id: { type: integer }
        userId: { type: string }
        email: { type: string }
        roleId: { type: integer }
        status: { type: string }
        createdAt: { type: string }
        roleName: { type: string }
        roleSlug: { type: string }
      required: [id, userId, email, roleId, status, createdAt, roleName, roleSlug]
...